+++ packages/plugin-inquirer-cli/CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+### Added
+- feat: Add emojis option
+
 ### Removed
 - feat: BREAKING CHANGE. Remove support for v1 legacy mode